provider/google: Refactor safeRetry into type safe class.
